prevent recursion with known urls
